// frontend/src/api/generated/models/salesPOLineResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface SalesPOLineResponse {
  id: string;
  sales_po_id: string;
  product_variant_id: string;
  quantity: number;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  unit_price: string;
  fulfilled_qty: number;
}